+++ src/dmpack.f90
@@ -82,6 +82,7 @@ module dmpack
     use :: dm_ipc
     use :: dm_ipc_message
     use :: dm_ipc_mutex
+    use :: dm_ipc_thread
     use :: dm_job
     use :: dm_job_list
     use :: dm_js